// src/interfaces/formulario.interface.ts
export interface Formulario {
    nombre?: any;
    paterno?: any;
    materno?: any;
    nacimiento?: any;
    curp?: any;
    sexo?: any;
    estado_civil?: any;
    calle?:any;
    num_interior?: any;
    colonia?: any;
    cp?: any;
    ciudad?: any;
    municipio?: any;
    estado?: any;
    tel?: any;
    matricula?: any;
    celular?: any;
    tel_paterno?: any;
    facebook?: any;
    twitter?: any;
    email?: any;
    carrera?: any;
    especialidad?: any;
    egreso?: any;
    titulado?: any;
    ingles?: any;
    idioma_porcentaje?: any;
    idioma_nombre?: any;
    softwares?: any;
    docentes?: any;
    plan_estudios?: any;
    investigacion_desarrollo?: any;
    investigacion_enseñanza?: any;
    infraestructura?: any;
    residencia?: any;

}